+++ mvc/src/main/java/nextstep/mvc/controller/asis/HandlerExecutionAdapter.java
@@ -1,7 +1,8 @@
-package nextstep.mvc;
+package nextstep.mvc.controller.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
