+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerControllerAdapter.java
@@ -1,7 +1,8 @@
-package nextstep.mvc;
+package nextstep.mvc.controller.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
